Handle NULL data in plots.
git-svn-id: https://svn.broadinstitute.org/CellProfiler/trunk/CPAnalyst@8724 1fc53939-2000-0410-845c-e8453a809027

+++ src/Density.py
@@ -218,7 +218,7 @@ class DensityPanel(PlotPanel):
         self.subplot = self.figure.add_subplot(111)
             
         for i, pt_list in enumerate(self.point_lists):
-            plot_pts = np.array(pt_list)
+            plot_pts = np.array(pt_list).astype(float)
             
             if self.x_scale == LOG_SCALE:
                 plot_pts = plot_pts[(plot_pts[:,0]>0)]
